feat(recipe): Code clean up
- Rename stage2 -> parser
- Remove stage1
- Remove unused code
